応募APIを修正

+++ app/Services/ApplyService.php
@@ -4,12 +4,13 @@ namespace App\Services;
 
 use App\Repositories\ApplyRepository;
 use App\Repositories\CompanyRepository;
+use App\Repositories\UserRepository;
 use Exception;
 use Illuminate\Support\Facades\Log;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Validation\Rule;
 
-class ApplyService
+class ApplyService extends Service
 {
     /** @var \App\Repositories\ApplyRepository */
     private $applyRepository;
@@ -17,38 +18,63 @@ class ApplyService
     /** @var \App\Repositories\CompanyRepository */
     private $companyRepository;
 
+    /** @var \App\Repositories\UserRepository */
+    private $userRepository;
+
     public function __construct(
         ApplyRepository $applyRepository,
         CompanyRepository $companyRepository,
+        UserRepository $userRepository,
     ) {
         $this->applyRepository = $applyRepository;
         $this->companyRepository = $companyRepository;
+        $this->userRepository = $userRepository;
     }
 
-    public function index(int $userId): \Illuminate\Http\JsonResponse
+    public function index(int $userId): \Illuminate\Database\Eloquent\Collection|array
     {
         try {
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
+
             $applies = $this->applyRepository->getBy(['user_id' => $userId]);
 
-            return response()->ok($applies);
+            return $applies;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
-    public function show(int $userId, int $applyId): \Illuminate\Http\JsonResponse
+    public function show(int $userId, int $applyId): \App\Models\Apply|array
     {
-        $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
+        try {
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
-        if ( $apply === null ) {
-            return response()->notFound();
+            $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
+            if ( $apply === null ) {
+                Log::error( __METHOD__ . ": Apply not found. (user_id={$userId}, apply_id={$applyId})" );
+                return $this->errorApplyNotFound();
+            }
+
+            return $apply;
+
+        } catch ( Exception $e ) {
+            Log::error(__METHOD__);
+            Log::error($e);
+
+            return $this->errorInternalServerError();
         }
-
-        return response()->ok($apply);
     }
 
     public function validateStore(array $postedParams): bool|array
@@ -64,36 +90,40 @@ class ApplyService
         $validator->setAttributeNames(['status' => '選考ステータス']);
 
         if ( $validator->fails() ) {
-            return ['errors' => $validator->errors()->getMessages()];
+            return $this->errorBadRequest( $validator->errors()->getMessages() );
         }
 
         return true;
     }
 
-    public function store(int $userId, array $postedParams): \Illuminate\Http\JsonResponse
+    public function store(int $userId, array $postedParams): \App\Models\Apply|array
     {
         try {
-            $companyId = $postedParams['company_id'];
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
             // 適切な企業か確認
+            $companyId = $postedParams['company_id'];
             $company = $this->companyRepository->findBy(['user_id' => $userId, 'company_id' => $companyId]);
-
             if ( $company === null ) {
-                Log::error( __METHOD__ . ": company not found. (company_id={$companyId}, user_id={$userId})");
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Company not found. (user_id={$userId}, company_id={$companyId})" );
+                return $this->errorCompanyNotFound();
             }
 
             $params = array_merge(['user_id' => $userId], $postedParams);
 
             $apply = $this->applyRepository->create($params);
 
-            return response()->ok($apply->fresh());
+            return $apply;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
@@ -109,61 +139,71 @@ class ApplyService
         $validator->setAttributeNames(['status' => '選考ステータス']);
 
         if ( $validator->fails() ) {
-            return ['errors' => $validator->errors()->getMessages()];
+            return $this->errorBadRequest( $validator->errors()->getMessages() );
         }
 
         return true;
     }
 
-    public function update(int $userId, int $applyId, array $postedParams): \Illuminate\Http\JsonResponse
+    public function update(int $userId, int $applyId, array $postedParams): \App\Models\Apply|array
     {
         try {
-            $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
+            $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
             if ( $apply === null ) {
-                Log::error( __METHOD__ . ": apply not found. (apply_id={$applyId}, user_id={$userId})");
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Apply not found. (user_id={$userId}, apply_id={$applyId})" );
+                return $this->errorApplyNotFound();
             }
 
             $isSuccess = $this->applyRepository->update($apply, $postedParams);
 
             if ( ! $isSuccess ) {
-                throw new Exception( __METHOD__ . ": Failed update apply. (apply_id={$applyId})");
+                throw new Exception( __METHOD__ . ": Failed update apply. (apply_id={$applyId}, user_id={$userId}, posted_params=" . json_encode($postedParams, JSON_UNESCAPED_UNICODE) . ")");
             }
 
-            return response()->ok($apply->fresh());
+            return $apply->fresh();
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 
-    public function delete(int $userId, int $applyId): \Illuminate\Http\JsonResponse
+    public function delete(int $userId, int $applyId): \App\Models\Apply|array
     {
         try {
-            $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
+            $user = $this->userRepository->find($userId);
+            if ( $user === null ) {
+                Log::error( __METHOD__ . ": User not found. (user_id={$userId})" );
+                return $this->errorUserNotFound();
+            }
 
+            $apply = $this->applyRepository->findBy(['user_id' => $userId, 'apply_id' => $applyId]);
             if ( $apply === null ) {
-                Log::error( __METHOD__ . ": apply not found. (apply_id={$applyId}, user_id={$userId})");
-                return response()->notFound();
+                Log::error( __METHOD__ . ": Apply not found. (user_id={$userId}, apply_id={$applyId})" );
+                return $this->errorApplyNotFound();
             }
 
             $isSuccess = $this->applyRepository->delete($apply);
 
             if ( ! $isSuccess ) {
-                throw new Exception( __METHOD__ . ": Failed delete apply. (apply_id={$applyId})");
+                throw new Exception( __METHOD__ . ": Failed delete apply. (user_id={$userId}, apply_id={$applyId})");
             }
 
-            return response()->ok($apply);
+            return $apply;
 
         } catch ( Exception $e ) {
             Log::error(__METHOD__);
             Log::error($e);
 
-            return response()->internalServerError();
+            return $this->errorInternalServerError();
         }
     }
 }
